Refactor RadarGrid
